requirements.txt
django
djangorestframework
requests
selenium